docs,demo: showcase draining progress-bar color
- Move progress-color toast into OverlaySection (reachable Modal tab); revert orphaned FeedbackSection
- Add live toast-progress-color playground example + docs section
- Document progress object/function color in the props table
