+++ apps/examples/kit-demo/src/sections/FeedbackSection.vue
@@ -1,14 +1,6 @@
 <script setup lang="ts">
-import { ref } from 'vue'
-import { ToastProvider, ToastViewport } from '@vyui/core'
-import { VyAlert, VyButton, VyToast } from '@vyui/kit'
-
-// Live progress demo — flip the key to remount a fresh toast so its
-// auto-dismiss timer (and the draining progress bar) restart from full.
-const progressToast = ref(0)
-function showProgressToast() {
-  progressToast.value += 1
-}
+import { ToastProvider } from '@vyui/core'
+import { VyAlert, VyToast } from '@vyui/kit'
 </script>
 
 <template>
@@ -64,28 +56,6 @@ function showProgressToast() {
           icon="icon-park-outline:close-one"
         />
       </view>
-    </ToastProvider>
-
-    <!-- Runtime toast whose countdown bar recolors as it drains. `progress.color`
-         takes a function of the remaining fraction (1 → 0), so the bar goes
-         green → amber → red on the way out. Tapping remounts (via `:key`) so it
-         restarts from full. -->
-    <view class="flex flex-col gap-3">
-      <text class="text-slate-900 text-base font-semibold px-1">Toast progress color</text>
-      <VyButton label="Show progress toast" color="neutral" variant="outline" @tap="showProgressToast" />
-    </view>
-    <ToastProvider v-if="progressToast" :duration="6000">
-      <ToastViewport position="top" :style="{ top: '60px', zIndex: 60 }">
-        <VyToast
-          :key="progressToast"
-          color="neutral"
-          title="Saving changes…"
-          description="This closes on its own — watch the bar go green → amber → red."
-          icon="icon-park-outline:check-one"
-          :progress="{ color: (p) => p > 0.5 ? 'success' : p > 0.25 ? 'warning' : 'error' }"
-          :close="false"
-        />
-      </ToastViewport>
     </ToastProvider>
   </view>
 </template>
